Fix a bug when player leaving on last turn causes game to be stuck.
If player's turn is last in the game, and there's still other players
remain in-game, then leaving will not properly finish the game and turn
will switch to next player, even if he has finished his last turn
already.

This causes game to be stuck, because player has no valid moves and can
only leave the game.

This fix adds an extra check for board completion after zeroing the
leaving player's board and causes stop_game to be properly called.

Signed-off-by: Vitaly Ostrosablin <tmp6154@yandex.ru>