Add a /kick command to allow owner to kick player, whose turn it is now.
Imagine a situation where a single player drops the game, which is in
progress without leaving. Unless a game is stopped - this will be a
dead-end situation, since permanently idling player would prevent
everyone from progressing. This patch introduces /kick command that
allows owner to kick idling players in that situation.

Signed-off-by: Vitaly Ostrosablin <tmp6154@yandex.ru>