Prohibit starting new game and stopping game for non-owner after timeout
This was an archaic hack, made when there was no option to leave/kick
while a game in progress. It was possible for a person to stale an active
game within chat by idling in their turn. And even to start a
single-player game and stay idle, effectively locking out everyone in chat
from game.

That's why idle timeout was introduced - after some time spent idling within
one's turn, any person, even non-owner could stop a stale game or start a new
one.

But later, when leave and kick were introduced, those become effectively
obsoleted, because now after timeout, any person in chat (including those
not in game) could kick the idling player, even if idler is owner of game.

Therefore, it's more logical to drop legacy commands now (they are too much
overpowered), and just give a temporary ability to kick idler after
timeout. This is much more preferred, because even if idler wants to spoil
game for everyone, there might be still people who are willing to continue
the game and just kick the idler.